refactoring fixtures for less imports
